requirements.txt
flask>=2.3.0
numpy>=1.24.0
pillow>=10.0.0
tensorflow>=2.14.0
scikit-learn>=1.3.0
werkzeug>=2.3.0
gunicorn>=21.0.0